attendance_system_v3/backend/requirements.txt
flask
flask-cors
mysql-connector-python
numpy